+++ ios/LastSingularity/LastSingularityApp.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct LastSingularityApp: App {
+    var body: some Scene {
+        WindowGroup {
+            WebGameView()
+                .ignoresSafeArea()
+                .persistentSystemOverlays(.hidden)
+        }
+    }
+}